Lösch-Funktion für jede Tabelle - Fortsetzung

+++ notendb/cl_instrument.php
@@ -7,6 +7,8 @@
   public $ID;
   public $Name;
   public $titles_selected_list; 
+  public $Title='Instrument';
+  public $Titles='Instrumente';  
 
   public function __construct(){
     $this->table_name='instrument'; 
@@ -165,22 +167,34 @@
     include_once("cl_db.php");
     $conn = new DbConn(); 
     $db=$conn->db; 
+
+    $select = $db->prepare("SELECT * from satz_schwierigkeitsgrad WHERE InstrumentID=:InstrumentID");
+    $select->bindValue(':InstrumentID', $this->ID); 
+    $select->execute();  
+    if ($select->rowCount() > 0 ){
+      $this->load_row(); 
+      echo '<p>Das Instrument ID '.$this->ID.' "'.$this->Name.'" 
+        kann nicht gelöscht werden, da noch eine Zuordnung auf '.$select->rowCount().' 
+        Sätze existiert. </p>';   
+      return false;            
+    }
  
     $delete = $db->prepare("DELETE FROM `instrument` WHERE ID=:ID"); 
     $delete->bindValue(':ID', $this->ID);  
 
     try {
       $delete->execute(); 
-      echo '<p>Die Zeile wurde gelöscht.</p>';          
+      echo '<p>Die Zeile wurde gelöscht.</p>'; 
+      return true;         
     }
     catch (PDOException $e) {
       include_once("cl_html_info.php"); 
       $info = new HtmlInfo();      
       $info->print_user_error(); 
       $info->print_error($delete, $e);  
+      return false;  
     }  
-  }  
-
+  } 
 
 }
